Create synchronous-radioactive-cryptocurrency-simulation.nb
This is a simulation of cryptocurrency that has a half-life. The goal is to make a currency that allows taxation without theft.  This takes the idea from radioactive decay in some elements, where instead of an atom, it is a coin.

Ideally, the next step would be use the Actor Model, where each coin is an actor. It keeps its lifespan private, and it securely randomly exchanges lifespans with other currencies so it is unknown what a particular coins lifespan is.